Jeg har invertert noen dependencies, så ikke UtbetalingService har en avhengighet til IverksettClient, men kun til et interface i eget namespace. Tanken er å eksemplifisere ports and adapters arkitekturen
