+++ src/main/kotlin/no/nav/tiltakspenger/utbetaling/domene/GrunnlagDTO.kt
@@ -1,0 +1,9 @@
+package no.nav.tiltakspenger.utbetaling.domene
+
+import java.time.LocalDate
+
+data class GrunnlagDTO(
+    val behandlingId: BehandlingId,
+    val fom: LocalDate,
+    val tom: LocalDate,
+)
