+++ src/test/kotlin/no/nav/tiltakspenger/utbetaling/service/UtbetalingServiceTest.kt
@@ -6,6 +6,7 @@ import no.nav.dagpenger.kontrakter.felles.GeneriskIdSomUUID
 import no.nav.dagpenger.kontrakter.felles.StønadTypeTiltakspenger
 import no.nav.dagpenger.kontrakter.iverksett.StønadsdataTiltakspengerDto
 import no.nav.dagpenger.kontrakter.iverksett.UtbetalingDto
+import no.nav.tiltakspenger.utbetaling.client.iverksett.IverksettMapper.mapIverksettDTO
 import no.nav.tiltakspenger.utbetaling.domene.SakId
 import no.nav.tiltakspenger.utbetaling.domene.Satser
 import no.nav.tiltakspenger.utbetaling.domene.TiltakType
@@ -62,17 +63,83 @@ class UtbetalingServiceTest {
             saksbehandler = "saksbehandler",
             beslutter = "beslutter",
             utbetalinger = listOf(
-                UtbetalingDag(dato = 1.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 2.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.DelvisUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 3.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.DelvisUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 4.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 5.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 6.januar(), tiltaktype = TiltakType.GRUPPEAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 7.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 8.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 9.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 10.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 11.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
+                UtbetalingDag(
+                    dato = 1.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 2.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.DelvisUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 3.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.DelvisUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 4.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 5.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 6.januar(),
+                    tiltaktype = TiltakType.GRUPPEAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 7.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 8.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 9.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 10.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 11.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
             ),
             forrigeVedtak = null,
         )
@@ -82,12 +149,42 @@ class UtbetalingServiceTest {
         }
 
         dto.vedtak.utbetalinger shouldContainExactlyInAnyOrder listOf(
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 1.januar(), tilOgMedDato = 1.januar(), stønadsdata = enkeltAmoUtenBarn()),
-            UtbetalingDto(beløpPerDag = sats.satsDelvis, fraOgMedDato = 2.januar(), tilOgMedDato = 3.januar(), stønadsdata = enkeltAmoUtenBarn()),
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 4.januar(), tilOgMedDato = 5.januar(), stønadsdata = enkeltAmoUtenBarn()),
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 6.januar(), tilOgMedDato = 6.januar(), stønadsdata = gruppeAmoUtenBarn()),
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 7.januar(), tilOgMedDato = 8.januar(), stønadsdata = enkeltAmoUtenBarn()),
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 11.januar(), tilOgMedDato = 11.januar(), stønadsdata = enkeltAmoUtenBarn()),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 1.januar(),
+                tilOgMedDato = 1.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.satsDelvis,
+                fraOgMedDato = 2.januar(),
+                tilOgMedDato = 3.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 4.januar(),
+                tilOgMedDato = 5.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 6.januar(),
+                tilOgMedDato = 6.januar(),
+                stønadsdata = gruppeAmoUtenBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 7.januar(),
+                tilOgMedDato = 8.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 11.januar(),
+                tilOgMedDato = 11.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
         )
     }
 
@@ -107,17 +204,83 @@ class UtbetalingServiceTest {
             saksbehandler = "saksbehandler",
             beslutter = "beslutter",
             utbetalinger = listOf(
-                UtbetalingDag(dato = 1.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 2.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.DelvisUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 3.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.DelvisUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 4.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 5.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 6.januar(), tiltaktype = TiltakType.GRUPPEAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 7.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 8.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 9.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 10.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
-                UtbetalingDag(dato = 11.januar(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = løpenr),
+                UtbetalingDag(
+                    dato = 1.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 2.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.DelvisUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 3.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.DelvisUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 4.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 5.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 6.januar(),
+                    tiltaktype = TiltakType.GRUPPEAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 7.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 8.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 9.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 10.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
+                UtbetalingDag(
+                    dato = 11.januar(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = løpenr,
+                ),
             ),
             forrigeVedtak = null,
         )
@@ -127,19 +290,79 @@ class UtbetalingServiceTest {
         }
 
         dto.vedtak.utbetalinger shouldContainExactlyInAnyOrder listOf(
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 1.januar(), tilOgMedDato = 1.januar(), stønadsdata = enkeltAmoUtenBarn()),
-            UtbetalingDto(beløpPerDag = sats.satsDelvis, fraOgMedDato = 2.januar(), tilOgMedDato = 3.januar(), stønadsdata = enkeltAmoUtenBarn()),
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 4.januar(), tilOgMedDato = 5.januar(), stønadsdata = enkeltAmoUtenBarn()),
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 6.januar(), tilOgMedDato = 6.januar(), stønadsdata = gruppeAmoUtenBarn()),
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 7.januar(), tilOgMedDato = 8.januar(), stønadsdata = enkeltAmoUtenBarn()),
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 11.januar(), tilOgMedDato = 11.januar(), stønadsdata = enkeltAmoUtenBarn()),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 1.januar(),
+                tilOgMedDato = 1.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.satsDelvis,
+                fraOgMedDato = 2.januar(),
+                tilOgMedDato = 3.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 4.januar(),
+                tilOgMedDato = 5.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 6.januar(),
+                tilOgMedDato = 6.januar(),
+                stønadsdata = gruppeAmoUtenBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 7.januar(),
+                tilOgMedDato = 8.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 11.januar(),
+                tilOgMedDato = 11.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
 
-            UtbetalingDto(beløpPerDag = sats.satsBarnetillegg * 2, fraOgMedDato = 1.januar(), tilOgMedDato = 1.januar(), stønadsdata = enkeltAmoMedBarn()),
-            UtbetalingDto(beløpPerDag = sats.satsBarnetilleggDelvis * 2, fraOgMedDato = 2.januar(), tilOgMedDato = 3.januar(), stønadsdata = enkeltAmoMedBarn()),
-            UtbetalingDto(beløpPerDag = sats.satsBarnetillegg * 2, fraOgMedDato = 4.januar(), tilOgMedDato = 5.januar(), stønadsdata = enkeltAmoMedBarn()),
-            UtbetalingDto(beløpPerDag = sats.satsBarnetillegg * 2, fraOgMedDato = 6.januar(), tilOgMedDato = 6.januar(), stønadsdata = gruppeAmoMedBarn()),
-            UtbetalingDto(beløpPerDag = sats.satsBarnetillegg * 2, fraOgMedDato = 7.januar(), tilOgMedDato = 8.januar(), stønadsdata = enkeltAmoMedBarn()),
-            UtbetalingDto(beløpPerDag = sats.satsBarnetillegg * 2, fraOgMedDato = 11.januar(), tilOgMedDato = 11.januar(), stønadsdata = enkeltAmoMedBarn()),
+            UtbetalingDto(
+                beløpPerDag = sats.satsBarnetillegg * 2,
+                fraOgMedDato = 1.januar(),
+                tilOgMedDato = 1.januar(),
+                stønadsdata = enkeltAmoMedBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.satsBarnetilleggDelvis * 2,
+                fraOgMedDato = 2.januar(),
+                tilOgMedDato = 3.januar(),
+                stønadsdata = enkeltAmoMedBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.satsBarnetillegg * 2,
+                fraOgMedDato = 4.januar(),
+                tilOgMedDato = 5.januar(),
+                stønadsdata = enkeltAmoMedBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.satsBarnetillegg * 2,
+                fraOgMedDato = 6.januar(),
+                tilOgMedDato = 6.januar(),
+                stønadsdata = gruppeAmoMedBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.satsBarnetillegg * 2,
+                fraOgMedDato = 7.januar(),
+                tilOgMedDato = 8.januar(),
+                stønadsdata = enkeltAmoMedBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.satsBarnetillegg * 2,
+                fraOgMedDato = 11.januar(),
+                tilOgMedDato = 11.januar(),
+                stønadsdata = enkeltAmoMedBarn(),
+            ),
         )
     }
 
@@ -160,35 +383,203 @@ class UtbetalingServiceTest {
             saksbehandler = "saksbehandler",
             beslutter = "beslutter",
             utbetalinger = listOf(
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = 1),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 1),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 1),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 1),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 1),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = 1),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = 1),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = 1),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 1),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 1),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 1),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 1),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = 1),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = 1),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 1,
+                ),
 
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = 2),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 2),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 2),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 2),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 2),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = 2),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = 2),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = 2),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 2),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 2),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 2),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.FullUtbetaling, meldekortId = meldekortId, løpenr = 2),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = 2),
-                UtbetalingDag(dato = dato.hent(), tiltaktype = TiltakType.ENKELAMO, status = UtbetalingDagStatus.IngenUtbetaling, meldekortId = meldekortId, løpenr = 2),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.FullUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
+                UtbetalingDag(
+                    dato = dato.hent(),
+                    tiltaktype = TiltakType.ENKELAMO,
+                    status = UtbetalingDagStatus.IngenUtbetaling,
+                    meldekortId = meldekortId,
+                    løpenr = 2,
+                ),
             ),
             forrigeVedtak = null,
         )
@@ -198,10 +589,30 @@ class UtbetalingServiceTest {
         }
 
         dto.vedtak.utbetalinger shouldContainExactlyInAnyOrder listOf(
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 2.januar(), tilOgMedDato = 5.januar(), stønadsdata = enkeltAmoUtenBarn()),
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 9.januar(), tilOgMedDato = 12.januar(), stønadsdata = enkeltAmoUtenBarn()),
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 16.januar(), tilOgMedDato = 19.januar(), stønadsdata = enkeltAmoUtenBarn()),
-            UtbetalingDto(beløpPerDag = sats.sats, fraOgMedDato = 23.januar(), tilOgMedDato = 26.januar(), stønadsdata = enkeltAmoUtenBarn()),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 2.januar(),
+                tilOgMedDato = 5.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 9.januar(),
+                tilOgMedDato = 12.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 16.januar(),
+                tilOgMedDato = 19.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
+            UtbetalingDto(
+                beløpPerDag = sats.sats,
+                fraOgMedDato = 23.januar(),
+                tilOgMedDato = 26.januar(),
+                stønadsdata = enkeltAmoUtenBarn(),
+            ),
         )
     }
 }
@@ -231,6 +642,7 @@ private fun enkeltAmoUtenBarn() =
         stønadstype = StønadTypeTiltakspenger.ENKELTPLASS_AMO,
         barnetillegg = false,
     )
+
 private fun enkeltAmoMedBarn() =
     StønadsdataTiltakspengerDto(
         stønadstype = StønadTypeTiltakspenger.ENKELTPLASS_AMO,
